Leave options as they were
