+++ table/options/options.go
@@ -8,8 +8,6 @@ import (
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/allocator"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/types"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/value"
-	"github.com/ydb-platform/ydb-go-sdk/v3/retry"
-	"github.com/ydb-platform/ydb-go-sdk/v3/trace"
 )
 
 func WithShardKeyBounds() DescribeTableOption {
@@ -871,28 +869,6 @@ func WithKeepInCache(keepInCache bool) ExecuteDataQueryOption {
 	)
 }
 
-type TableOptions struct {
-	Label           string
-	Idempotent      bool
-	TxSettings      *TransactionSettings
-	TxCommitOptions []CommitTransactionOption
-	RetryOptions    []retry.Option
-	Trace           *trace.Table
-	CallOptions     []grpc.CallOption
-}
-
-type TransactionSettings struct {
-	settings Ydb_Table.TransactionSettings
-}
-
-func (t *TransactionSettings) Settings() *Ydb_Table.TransactionSettings {
-	if t == nil {
-		return nil
-	}
-
-	return &t.settings
-}
-
 type withCallOptions []grpc.CallOption
 
 func (opts withCallOptions) ApplyExecuteScanQueryOption(d *ExecuteScanQueryDesc) []grpc.CallOption {
@@ -907,10 +883,6 @@ func (opts withCallOptions) ApplyExecuteDataQueryOption(
 	d *ExecuteDataQueryDesc, a *allocator.Allocator,
 ) []grpc.CallOption {
 	return opts
-}
-
-func (opts withCallOptions) ApplyTableOption(tableOpts *TableOptions) {
-	tableOpts.CallOptions = append(tableOpts.CallOptions, opts...)
 }
 
 // WithCallOptions appends flag of commit transaction with executing query
